src/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.7.7
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.8.1
astroid==3.3.8
attrs==24.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0 
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cohere==4.37
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
docarray==0.39.1
durationpy==0.9
effdet==0.4.1
elastic-transport==8.17.0
elasticsearch==8.17.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastavro==1.10.0
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.9.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
gradio==5.13.1
gradio_client==1.6.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kagglehub==0.3.6
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-elasticsearch==0.3.2
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-openai==0.3.2
langchain-pinecone==0.2.2
langchain-tests==0.3.9
langchain-text-splitters==0.3.5
langdetect==1.0.9
langserve==0.3.1
langsmith==0.3.1
lark==1.2.2
layoutparser==0.3.4
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
ndjson==0.3.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.6
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.60.1
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
panns-inference==0.1.1
patsy==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pikepdf==9.5.1
pillow==11.1.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pooch==1.8.2
portalocker==3.1.1
posthog==3.8.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pylint==3.3.3
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.4.0
setuptools==75.8.0
shellingham==1.5.4
simsimd==6.2.1
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.37
sse-starlette==2.2.1
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.1
syrupy==4.8.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchlibrosa==0.1.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.13
unstructured-client==0.29.0
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
wikipedia==1.4.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
yt-dlp==2025.1.26
zstandard==0.23.0
